### 2021-02-25 #3063 @whedon
[PRE REVIEW]: StatSailr: Reproducible and Easy Statistics Environment with StatSailr Script
**Submitting author:** @niceume (<a href="http://orcid.org/0000-0001-7015-621X">Toshihiro Umehara</a>)
**Repository:** <a href="https://github.com/niceume/statsailr/" target ="_blank">https://github.com/niceume/statsailr/</a>
**Version:** v0.7.2
