### 2021-02-25 #3063 @kthyng
Hi @niceume and thanks for your submission! I want to be sure about the fit of your package within the scope of JOSS, so I am going to ping the editorial board to take a look with the scope query label. Someone will get back to you after this in a week or two.
