### 2021-02-26 #3063 @arfon
> @arfon Did something get messed up in here? I see a "juliacon" in the error dump above.

Lemme dig into this. I think there's some kind of issue on the Whedon bot server side.
