### 2021-03-01 #3063 @arfon
@niceume - this looks like a useful Ruby library. After a quick inspection of the code, it looks like there aren't any automated tests or testing procedures here to verify the functionality of the software? The JOSS review [includes specific criteria for verifying the testing approach](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) - could you explain how you would expect a reviewer to test your software?
