### 2021-03-01 #3063 @niceume
@arfon Thank you for having a look at my Ruby program. This gem installs CLI programs called sailr and sailrREPL. Actually, how to test is not mentioned, but I would like reviewers to run examples described in README.

This gem requires some setups to run. If Docker is allowed, please have a look at the bottom of this page. https://statsailr.io/install/install.html
Also, how it works can be seen here. https://statsailr.io/demo/demo.html
