### 2021-03-06 #3063 @niceume
@danielskatz - I am sorry for being late. I needed to take time for my family. Testing part is now added, and it can be done with docker.
