### 2021-03-27 #3063 @niceume
@arfon - Assigned editor seems to be busy now. Can I ask another editor for reviewing? This Ruby program converts StatSailr script into R's internal expression and executes it. I see the reviewer list, and I would like to ask lutzhamel to review it.
