### 2021-03-28 #3063 @arfon
> I'm pretty sure `@lorenanicole` is currently unable to take on an additional paper so I'm going to make this submission as waitlisted for now.

@niceume - as per my comment two weeks ago, your submission is currently waitlisted (there is no editor assigned currently to handle it). Once an editor is assigned, they will then start the process of identifying reviewers.

JOSS is very busy right now. Thanks for your patience.
