### 2021-04-06 #3063 @danielskatz
@fboehm - is there any chance you could take this on as well?  It's related, but Ruby.  It would be useful for the same editor to handle both if possible
