### 2021-04-16 #3170 @niceume
@fboehm @g4brielvs - Thank you for your reviewing. I come to think that I need to once withdraw this StatSailr's submission. I have another repository reviewed at https://github.com/openjournals/joss-reviews/issues/3166#event-4590901219 , and I am receiving comments. Considering those comments, I also think I need to update this StatSailr's paper and README especially about testing part.

I am sorry for taking your time, but I need to update this paper and README. After changes, I will re-submit this paper in the future.
