### 2021-04-16 #3170 @nachootal
Hi @niceume , thanks for letting us know. I'll wait for you to let us know when can we continue. As I had already started reviewing it, I have a suggestion. I understand the focus of the software is skilled scientists, specifically staticians, that might very well know how to install R and have it set up on their computers but newbies might not be so skilled. Since this is an important requirement to your software, I'd add a link to the official installation documentation of R.
